+++ CHANGELOG.md
@@ -20,6 +20,7 @@
     - Parser generates less virtual nodes, so parse->stringify is much closer to original code (white spaces and comments mostly lost)
     - Suggestion subsystem is moved from parser to separate module which is using in stat mode only
     - Various fixes and improvements in suggestions
+    - The new approach allows to implement more complex suggestion scenarios like suggestions in array values for operators `in`, `not in`, `has` and `has no` which was added (e.g. in query `["a", "b", 1, 2][$ in ["b", 2]]` jora will suggest only `"a"` and `1` values in array after `in` operator)
 - Changed `=` and `!=` operators to use `Object.is()` instead of `===` and `!==` operators
 
 ## 1.0.0-alpha.13 (January 6, 2020)